Moved json_symbol and json_symbol_table from util to the json-symtab-language subfolder
