+++ src/json-symtab-language/json_symbol_table.h
@@ -9,8 +9,8 @@ Author: Chris Smowton, chris.smowton@diffblue.com
 #ifndef CPROVER_UTIL_JSON_SYMBOL_TABLE_H
 #define CPROVER_UTIL_JSON_SYMBOL_TABLE_H
 
-#include "json.h"
-#include "symbol_table.h"
+#include <util/json.h>
+#include <util/symbol_table.h>
 
 void symbol_table_from_json(const jsont &, symbol_tablet &);
