+++ src/util/Makefile
@@ -39,9 +39,6 @@ SRC = arith_tools.cpp \
       json_expr.cpp \
       json_irep.cpp \
       json_stream.cpp \
-      json_symbol.cpp \
-      json_symbol_table.cpp \
-      json_symbol.cpp \
       lispexpr.cpp \
       lispirep.cpp \
       mathematical_types.cpp \
